lib: Add a new "IIOAttribute" trait.
This trait is used by the generic "read_attr" and "write_attr" methods in the
device and channel modules. It performs type-specific string-to-type
conversions and vice-versa. Adds an override for the bool type to respect the
fact that rust translates the true/false as "true"/"false" by default, whereas
the underlying sysfs expects "1"/"0", respectively.

It otherwise implements the traits for the other types that libiio contains
specialized functions for (f64 and i64).